feat(algorithms): majority element in array
Finds the majority element in the Array that occurs more than n/2 times using 3 different approaches:

1. Using sorting
2. Using a hash map
3. Using Boyer-Moore's Voting Algorithm

1. Using Sorting sorts the array in place and retrieves the element at n/2 index as the element
that occurs more than n/2 times in the array will be at this index

2. Using a hash map enables storing the count of each element in a map and iterates through
the key value pairs comparing the values against n/2 to find the element that has an occurrence of
more than n/2 and returns it

3. Using Boyer-Moor's voting algorithm assigns a potential majority candidate from the array keeping
track of a count while iterating through the array. If the current element being iterated is equal to
the candidate, the count is increment, otherwise, decremented. This will retain the candidate as it will
always have the lead over any other element. After the iteration, the majority element(candidate) is returned.